Refactor: Extract ResetSection component from ConnectView.tsx
- Moved ResetSection component and its props to app/client/connect/ResetSection.tsx.
- Exported ResetSectionProps for better reusability.
- Updated border color to use theme.palette.divider to follow design system guidelines.
- Cleaned up ConnectView.tsx imports and removed local component definition.
- Verified that all unit tests pass and UI remains consistent.
